Price the decoder header in stable u64 bit widths and drop the unused basis import (#2933 F14)
d03e19f90 priced each f64 header scalar with f64::BITS, which is unstable
on the pinned toolchain (E0658 float_bits_const), and imported
SaeBasisEvaluator although the evaluator call resolves through the trait
object. gam-sae did not compile under -D warnings. A persisted f64 is its
u64 bit pattern, so the header charge now uses u64::BITS, and the unused
import is removed. The charge is unchanged.

+++ crates/gam-sae/src/description_length.rs
@@ -16,7 +16,6 @@
 //! [`manifold_fit_description_length`].
 
 use crate::atom_codes::SparseAtomCodes;
-use crate::basis::SaeBasisEvaluator;
 use crate::manifold::SaeAtomGeometryPlan;
 use ndarray::ArrayView2;
 
@@ -678,7 +677,7 @@ pub fn persisted_decoder_dictionary_code(
     let plan_bytes = serde_json::to_vec(geometry_plans)
         .map_err(|error| format!("persisted decoder dictionary code: plan encoding failed: {error}"))?;
     let header_bits = f64::from(u8::BITS) * plan_bytes.len() as f64
-        + f64::from(f64::BITS) * (2 * k_atoms + output_side_scalars) as f64;
+        + f64::from(u64::BITS) * (2 * k_atoms + output_side_scalars) as f64;
     Ok(DictionaryCode::Quantized {
         blocks,
         header_bits,
